requirements.txt
flask==1.0.2
gunicorn==19.9.0
requests==2.19.1
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
oauth2client
